updated the rotate_card and some edits to make the XP bar work

+++ frontend/lib/Pages/service/database.dart
@@ -26,7 +26,7 @@ class UserModel {
   String name;
   int nutrition;
   String password;
-  int xp;
+  double xp;
 
   UserModel({
     this.accessory = 0,
